refactor: 💡 avoid duplicate type definition
