+++ src/globalState/scoreAtom.ts
@@ -1,7 +1,6 @@
 import { atomWithStorage } from "jotai/utils";
 import { useAtom } from "jotai/index";
 
-// type ScoreMap = [number, number, number, number];
 export type ScoreMap = number[];
 
 const STORAGE_KEY = "score";
